chore: rename some files [YTFRONT-4119]
