+++ packages/ui/src/ui/pages/tablet-errors-by-bundle/TabletErrorsByBundle.scss
@@ -1,4 +1,4 @@
-.yt-bundle-tablet-errors {
+.yt-tablet-errors-by-byndle {
     &__table {
         margin-bottom: 20px;
     }
