+++ packages/ui/src/ui/pages/tablet-errors-by-bundle/lazy.tsx
@@ -3,11 +3,11 @@ import React from 'react';
 import withLazyLoading from '../../hocs/withLazyLoading';
 
 function importPage() {
-    return import(/* webpackChunkName: "tablet-errors-by-bundle" */ './TabletErrors');
+    return import(/* webpackChunkName: "tablet-errors-by-bundle" */ './TabletErrorsByBundle');
 }
 
 export const TabletErrorsLazy = withLazyLoading(
     React.lazy(async () => {
-        return {default: (await importPage()).TabletErrors};
+        return {default: (await importPage()).TabletErrorsByBundle};
     }),
 );
